+++ src/main/java/site/scalarstudios/scalarpower/item/ScalarPowerCreativeTabs.java
@@ -50,8 +50,8 @@ public class ScalarPowerCreativeTabs {
             event.accept(ScalarPowerBlocks.INSULATED_COPPER_WIRE.get());
             event.accept(ScalarPowerBlocks.GOLD_WIRE.get());
             event.accept(ScalarPowerBlocks.INSULATED_GOLD_WIRE.get());
-            event.accept(ScalarPowerBlocks.GLASS_FIBER_WIRE.get());
-            event.accept(ScalarPowerBlocks.REINFORCED_GLASS_FIBER_WIRE.get());
+            event.accept(ScalarPowerBlocks.FIBER_GLASS_WIRE.get());
+            event.accept(ScalarPowerBlocks.REINFORCED_FIBER_GLASS_WIRE.get());
 
             /* Normal Blocks */
             event.accept(ScalarPowerBlocks.COBALT_BLOCK.get());